backend/pyproject.toml
[project]
name = "axio-backend"
version = "0.1.0"
description = "Evidence-grounded mediator assistant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pydantic-settings",
    "openai",
    "qdrant-client",
    "pdfplumber",
    "asyncpg",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py311"
